back/requirements.txt
requests~=2.25.1
fastapi~=0.111.1
pytest~=8.3.1
python-dotenv==1.0.0